updated 3/7

+++ api/Services/InvoiceService.cs
@@ -10,6 +10,7 @@ using api.Repository.Interfaces;
 using api.Services.Interfaces;
 using AutoMapper;
 using Microsoft.AspNetCore.Http.HttpResults;
+using Microsoft.EntityFrameworkCore;
 
 namespace api.Services
 {
@@ -17,10 +18,13 @@ namespace api.Services
     {
         private readonly IInvoiceRepository _invoicerepo;
         private readonly IMapper _mapper;
-        public InvoiceService(IInvoiceRepository invoicerepo, IMapper mapper)
+        private readonly AppDbContext _context;
+        public InvoiceService(AppDbContext context, IInvoiceRepository invoicerepo, IMapper mapper)
         {
             _invoicerepo = invoicerepo;
             _mapper = mapper;
+            _context = context;
+            
         }
         public async Task<InvoiceResponseDTO> CreateInvoiceAsync(CreateInvoiceMasterDTO dto)
         {
@@ -35,18 +39,34 @@ namespace api.Services
 
             }
 
-            var invoice = _mapper.Map<InvoiceMaster>(dto);
-            
-            foreach (var item in invoice.InvoiceItemDetails)
+            using var transaction = await _context.Database.BeginTransactionAsync();
+
+            try
             {
-                item.Total = item.Quantity * item.UnitPrice;
+
+                var invoice = _mapper.Map<InvoiceMaster>(dto);
+
+                foreach (var item in invoice.InvoiceItemDetails)
+
+                {
+                    item.Total = item.Quantity * item.UnitPrice;
+                }
+
+                invoice.TotalAmount = invoice.InvoiceItemDetails.Sum(i => i.Total);
+
+                var saved = await _invoicerepo.AddInvoiceAsync(invoice);
+
+                await transaction.CommitAsync();
+
+                return _mapper.Map<InvoiceResponseDTO>(saved);
+
             }
 
-            invoice.TotalAmount = invoice.InvoiceItemDetails.Sum(i => i.Total);
-
-            var saved = await _invoicerepo.AddInvoiceAsync(invoice);
-
-            return _mapper.Map<InvoiceResponseDTO>(saved);
+            catch
+            {
+                await transaction.RollbackAsync();
+                throw;
+            }
         }
 
         public async Task<List<InvoiceResponseDTO>> GetAllInvoiceAsync()
@@ -67,30 +87,63 @@ namespace api.Services
 
         public async Task<bool> UpdateInvoiceAsync(int id, UpdateInvoiceMasterDTO dto)
         {
-            var invoice = await _invoicerepo.GetInvoiceById(id);
+            using var transaction = await _context.Database.BeginTransactionAsync();
 
-            if (invoice == null) throw new NotFoundException("Invoice not found (Wrong id)");
-
-            if (string.IsNullOrWhiteSpace(dto.CustomerName))
+            try
             {
-                throw new BadRequestException("Customer name is required");
+
+                var invoice = await _invoicerepo.GetInvoiceById(id);
+
+                if (invoice == null) throw new NotFoundException("Invoice not found (Wrong id)");
+
+                if (string.IsNullOrWhiteSpace(dto.CustomerName))
+                {
+                    throw new BadRequestException("Customer name is required");
+                }
+                var updatedinvoice = _mapper.Map(dto, invoice);
+
+                updatedinvoice.Id = id;
+
+                foreach (var item in updatedinvoice.InvoiceItemDetails)
+                {
+                    item.Total = item.Quantity * item.UnitPrice;
+                }
+                updatedinvoice.TotalAmount = updatedinvoice.InvoiceItemDetails.Sum(i => i.Total);
+
+                var result = await _invoicerepo.UpdateInvoiceAsync(updatedinvoice);
+
+                await transaction.CommitAsync();
+
+                return result;
             }
-            var updatedinvoice = _mapper.Map(dto,invoice);
-
-            updatedinvoice.Id = id;
-
-            foreach (var item in updatedinvoice.InvoiceItemDetails)
+            catch
             {
-                item.Total = item.Quantity * item.UnitPrice;
+                await transaction.RollbackAsync();
+                throw;
             }
-            updatedinvoice.TotalAmount = updatedinvoice.InvoiceItemDetails.Sum(i => i.Total);
-
-            return await _invoicerepo.UpdateInvoiceAsync(updatedinvoice); 
         }
 
         public async Task<bool> DeleteInvoiceAsync(int id)
         {
-            return await _invoicerepo.DeleteInvoiceAsync(id);
+            using var transaction = await _context.Database.BeginTransactionAsync();
+            try
+            {
+                var invoice = await _invoicerepo.GetInvoiceById(id);
+                if (invoice == null)
+                {
+                    throw new NotFoundException("Invoice not found (Wrong Id)");
+                }
+                var deleted = await _invoicerepo.DeleteInvoiceAsync(id);
+
+                await transaction.CommitAsync();
+
+                return deleted;
+            }
+            catch
+            {
+                await transaction.RollbackAsync();
+                throw; 
+            }
         }
     }
 }
